+++ src/manager/manager.ts
@@ -1,3 +1,4 @@
+import {Iterable, List} from 'immutable';
 import {Observable} from 'rxjs/Observable';
 
 import {Injectable, Inject, Optional, Provider, OpaqueToken} from '@angular/core';
@@ -7,14 +8,13 @@ import {Codec} from 'caesium-core/codec';
 import {memoize} from 'caesium-core/decorators';
 import {ValueError} from 'caesium-core/exception';
 
-
-import {ModelMetadata} from '../model/metadata';
+import {ModelMetadata, PropertyMetadata, RefPropertyMetadata} from '../model/metadata';
 import {MetadataProvider} from '../model/metadata_provider';
 import {ModelFactory, createModelFactory} from '../model/factory';
-import {InvalidMetadata, ModelNotFoundException} from '../model/exceptions';
+import {InvalidMetadata, ModelNotFoundException, PropertyNotFoundException} from '../model/exceptions';
 import {ModelBase} from '../model/base';
 
-import {model, union, JsonObject} from '../json_codecs';
+import {model, itemList, union, JsonObject} from '../json_codecs';
 import {RequestFactory, Request} from './http';
 
 import {Search, SearchParameter,
@@ -22,7 +22,23 @@ import {Search, SearchParameter,
     SEARCH_PAGE_QUERY_PARAM, defaultSearchPageQueryParam
 } from './search';
 
-
+/**
+ * Provides a number of utility functions for models which meet the following criteria
+ *
+ * - The model must have a unique identifying property, with `key: true`
+ * - The model must have a path
+ * - Given an ID, the model can be loaded via a GET request to `${model.path}/${id}`
+ *    or return a 404 response if the ID was not found.
+ * - Given a list of IDs, the models can be loaded via a GET request to
+ *      `${model.path}?ids=${idList.join(',')}`
+ *   The request should return the results wrapped in a JSON object with a single key `items`
+ *   The result should _NOT_ be paginated.
+ *
+ * - An instance can be created via a POST request to `${model.path}`
+ * - An instance can be updated via a PUT request to `${model.path}/${model.id}`
+ *
+ * Only managed models can be used as reference properties.
+ */
 @Injectable()
 export class ModelManager {
     searchPageSize: number;
@@ -50,11 +66,25 @@ export class ModelManager {
 
     public getPath(type: Type/*<T>*/): Array<string> {
         let metadata = this.metadatas.for(type);
+
+        if (!metadata.isManaged)
+            throw new InvalidMetadata(`${type} is not a managed model`);
+
         return metadata.path;
     }
 
+    private getKey(managedModel: ModelBase): string | number | null {
+        // Also test whether the model is registered
+        let metadata = this.metadatas.for(managedModel);
 
-    getById<T extends ModelBase>(type: Type/*<T>*/, id: any, options?: {ignoreCache?: boolean}): Observable<T> {
+        if (!metadata.isManaged) {
+            throw new InvalidMetadata(`${metadata.type} is not a managed model`);
+        }
+
+        return managedModel.get(metadata.keyProperty.name);
+    }
+
+    load<T extends ModelBase>(type: Type/*<T>*/, id: string | number, options?: {ignoreCache?: boolean}): Observable<T> {
         let params: {[key: string]: string} = {};
 
         if (options && options.ignoreCache) {
@@ -65,6 +95,20 @@ export class ModelManager {
         return this.requests
             .get([...this.getPath(type), id.toString()], params)
             .send(this.getModelCodec<T>(type));
+    }
+
+    loadMany<T extends ModelBase>(type: Type/*<T>*/, ids: Iterable<any, string | number>, options?: {ignoreCache?: boolean}): Observable<List<T>> {
+        let params: {[key: string]: string} = {
+            'ids': ids.map(id => id.toString()).join(',')
+        };
+
+        if (options && options.ignoreCache) {
+            params['cache'] = `${new Date().valueOf()}`;
+        }
+
+        return this.requests
+            .get([...this.getPath(type)], params)
+            .send(itemList(this.getModelCodec<T>(type)));
     }
 
 
@@ -81,10 +125,10 @@ export class ModelManager {
         let metadata = this.metadatas.for(model);
         let codec = this.getModelCodec<T>(model);
 
-        if (model.id === null) {
-            request = this.requests.post([...metadata.path, 'create']);
+        if (this.getKey(model) === null) {
+            request = this.requests.post([...metadata.path]);
         } else {
-            request = this.requests.put([...metadata.path, `${model.id}`]);
+            request = this.requests.put([...metadata.path, this.getKey(model).toString()]);
         }
 
         return request.setRequestBody(model, codec).send<T>(codec);
@@ -99,6 +143,48 @@ export class ModelManager {
             this.searchPageSize,
             this.searchPageQueryParam
         );
+    }
+
+
+    /**
+     * Load the value of the reference property, given by either it's reference
+     * or property name, from the server, returning the model with the property
+     * resolved.
+     *
+     * @param type
+     * @param propNameOrRefName
+     * @returns {any}
+     */
+    resolve<T extends ModelBase>(model: T, propNameOrRefName:string):Observable<T> {
+        if (model.isResolved(propNameOrRefName)) {
+            return Observable.of(model);
+        }
+        // Cannot use this.metadatas, since the model may be nested and therefore not registered
+        let metadata: ModelMetadata = (model as any).__metadata__;
+        let property = metadata.getProperty(propNameOrRefName);
+
+        if (!property.isRef) {
+            return Observable.throw(new PropertyNotFoundException(propNameOrRefName, this, 'Reference'));
+        }
+        // Fetch the value of the foreign ID by name, even if the reference name was passed.
+        let idValue = model.get(property.name);
+
+
+        let prop = <RefPropertyMetadata>property;
+
+        if (isBlank(idValue)) {
+            // A null id maps to a null reference.
+            let resolved = model.set(prop.refName, null)
+            return Observable.of(resolved);
+        } else if (property.isMulti) {
+            return this.loadMany(prop.refType, idValue)
+                .map((foreignModels) => model.set(prop.refName, foreignModels))
+        } else {
+            return this.load(prop.refType, idValue)
+                .map((foreignModel) => {
+                    return model.set(prop.refName, foreignModel)
+                });
+        }
     }
 
 }