+++ src/model/base.ts
@@ -1,13 +1,13 @@
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
+import 'rxjs/add/observable/throw';
 
 
-import {List} from 'immutable';
+import {Iterable, List} from 'immutable';
 import {isDefined, isBlank, forEachOwnProperty} from 'caesium-core/lang';
 
 import {itemList} from '../json_codecs';
 import {ModelMetadata, BasePropertyMetadata, RefPropertyMetadata} from './metadata';
-import {ModelManager} from '../manager';
 
 import {copyModel, ModelConstructor} from './factory';
 import {initialModelValues, ModelValues, isModelValues, mutateModelValues} from './values';
@@ -15,14 +15,6 @@ import {initialModelValues, ModelValues, isModelValues, mutateModelValues} from 
 import {PropertyNotFoundException, ArgumentError} from './exceptions';
 
 export class ModelBase {
-    /**
-     * The server-assigned id of the model.
-     */
-    get id(): any {
-        // slight optimization
-        let idValue = this.__modelValues__.values.get('id');
-        return isDefined(idValue) ? idValue : null;
-    }
 
     // Do we really need to keep the metadata on the model?
     // Seems like it's exposing implementation details.
@@ -33,7 +25,7 @@ export class ModelBase {
         return (this.__metadata__.type as ModelConstructor<this>);
     }
 
-    constructor(id: number, ...args: any[]) {
+    constructor(...args: any[]) {
         // The arguments are passed in internally from the proxied type, ignoring the constructor.
         // The constructor is typed to satisfy the type checker, no matter
         // The /* true */ signature of the constructor is
@@ -44,26 +36,31 @@ export class ModelBase {
         // If modelValues is an Array, it will be the list of arguments
         // passed to the actual constructor
         // Otherwise, it's a ModelValues instance to use as __model_values__.
-        this.__metadata__ = id as any;
+        if (args.length < 2) {
+            // Should never happen
+            throw new ArgumentError('Invalid arguments for model constructor')
+        }
 
-        if (args.length > 0 && isModelValues(args[0])) {
-            this.__modelValues__ = args[0];
+        this.__metadata__ = args[0] as ModelMetadata;
+        this.__metadata__.checkValid();
+
+        let modelArgs = args[1];
+
+        if (modelArgs.length > 0 && isModelValues(modelArgs[0])) {
+            this.__modelValues__ = modelArgs[0];
         } else {
-            if (args.length !== this.__metadata__.properties.count()) {
-                throw new ArgumentError('Unexpected number of arguments');
+            if (modelArgs.length !== this.__metadata__.properties.count()) {
+                throw new ArgumentError(`Unexpected number of arguments for ${this.__metadata__.type}`);
             }
 
             this.__modelValues__ = mutateModelValues(initialModelValues, mutator => {
                 this.__metadata__.properties.valueSeq()
-                    .zip(List(args))
+                    .zip(List(modelArgs))
                     .forEach(([property, value]) => {
                         property.valueAccessor.set(mutator, value)
                     });
             });
         }
-
-        // Adds the property descriptors
-        //this.__metadata__.prepareInstance(this);
     }
 
     /**
@@ -125,30 +122,5 @@ export class ModelBase {
         return property.valueAccessor.has(this.__modelValues__, true);
     }
 
-    resolve(manager:ModelManager, propNameOrRefName:string):Observable<this> {
-        if (this.isResolved(propNameOrRefName)) {
-            return Observable.of(this);
-        }
-        let property = this.__metadata__.getProperty(propNameOrRefName);
-
-        if (!property.isRef) {
-            return Observable.throw(new PropertyNotFoundException(propNameOrRefName, this, 'Reference'));
-        }
-        var idValue = property.valueAccessor.get(
-            this.__modelValues__,
-            /* getRef: */ false
-        );
-
-        let refName = (<RefPropertyMetadata>property).refName;
-
-        if (isBlank(idValue)) {
-            // A null id maps to a null reference.
-            let resolved = this.set(refName, null)
-            return Observable.of(resolved);
-        }
-
-        return manager.getById(this.__metadata__.type, idValue)
-            .map((foreignModel) => this.set(refName, foreignModel));
-    }
 }