+++ test/manager/manager.spec.ts
@@ -2,13 +2,13 @@ import 'rxjs/add/operator/toPromise';
 import {List} from 'immutable';
 
 import {Inject, forwardRef, OpaqueToken} from '@angular/core';
-import {TestBed, inject} from '@angular/core/testing';
+import {TestBed, inject, async} from '@angular/core/testing';
 import {RequestMethod, HttpModule} from '@angular/http';
 
 import {Type} from 'caesium-core/lang';
 
-import {JsonObject, itemList} from '../../src/json_codecs';
-import {Model, ModelBase, RefProperty} from '../../src/model';
+import {JsonObject, num, itemList, model} from '../../src/json_codecs';
+import {Model, ModelBase, Property, RefProperty} from '../../src/model';
 import {Models} from '../../src/module';
 
 import {RequestFactory, ModelHttpModule} from '../../src/manager/http';
@@ -23,6 +23,12 @@ import {MockRequestFactory, MockRequest} from './request_factory.mock';
 
 @Model({kind: 'test::MyModel'})
 export class MyModel extends ModelBase {
+    constructor(
+        @Property('id', {key: true, codec: num})
+        public id: number
+    ) {
+        super(id);
+    }
 }
 
 
@@ -41,22 +47,24 @@ export class AbstractModelImpl2 extends AbstractModel {}
 @Model({kind: 'test::ReferencingModel'})
 class ReferencingModel extends ModelBase {
     constructor(
-        id: number,
-        @RefProperty('ref', {refName: 'ref', refType: MyModel})
-        refId: number
+        @RefProperty('refId', {refName: 'ref', refType: MyModel})
+        public refId: number
     ) {
-        super(id, refId);
+        super(refId);
     }
+
+    get ref(): MyModel { return this.get('ref'); }
 }
 
-describe('manager.base', () => {
+describe('manager.manager', () => {
     describe('ManagerBase', () => {
         beforeEach(() => {
             TestBed.configureTestingModule({
-                imports: [HttpModule,
+                imports: [
+                    HttpModule,
                     Models.provideMetadata([
                         MyModel,
-                        {type: AbstractModel, subtypes: [AbstractModelImpl1, AbstractModelImpl2]}
+                        {type: AbstractModel, subtypes: [AbstractModelImpl1, AbstractModelImpl2]},
                     ])
                 ],
                 providers: [
@@ -79,7 +87,7 @@ describe('manager.base', () => {
             expect(codec.encode(instance))
                 .toEqual({kind: 'test::MyModel', id: null});
 
-            expect(codec.decode({'kind': 'test::MyModel'}))
+            expect(codec.decode({'kind': 'test::MyModel', id: null}))
                 .toEqual(jasmine.any(MyModel));
         }));
 
@@ -87,17 +95,17 @@ describe('manager.base', () => {
             (manager: ModelManager) => {
                 var codec = manager.getModelCodec(AbstractModel);
 
-                let model1 = new AbstractModelImpl1(null);
-                let model2 = new AbstractModelImpl2(null);
+                let model1 = new AbstractModelImpl1();
+                let model2 = new AbstractModelImpl2();
 
                 expect(codec.encode(model1))
                     .toEqual(
-                        {kind: 'test::AbstractModelImpl1', id: null},
+                        {kind: 'test::AbstractModelImpl1'},
                         'should encode as AbstractModelImpl1'
                     );
                 expect(codec.encode(model2))
                     .toEqual(
-                        {kind: 'test::AbstractModelImpl2', id: null},
+                        {kind: 'test::AbstractModelImpl2'},
                         'should encode as AbstractModelImpl2'
                     );
 
@@ -111,7 +119,7 @@ describe('manager.base', () => {
         );
 
 
-        it('should be possible to get a model by id', inject(
+        it('should be possible to get a model by id', async(inject(
             [RequestFactory, ModelManager],
             (requestFactory: MockRequestFactory, manager: ModelManager) => {
                 requestFactory.sent$.subscribe((request: MockRequest)=> {
@@ -122,19 +130,47 @@ describe('manager.base', () => {
                     request.respond({status: 200, body: new MyModel(12345)});
                 });
 
-                manager.getById<MyModel>(MyModel, 12345).forEach(model => {
+                manager.load(MyModel, 12345).forEach(model => {
                     expect(model).toEqual(new MyModel(12345));
                 })
-            })
+            }))
         );
 
-        it('saving a model with null ID should POST to server', inject(
+        it('should be possible to gte a list of models by their IDs', async(inject(
+            [RequestFactory, ModelManager],
+            (requestFactory: MockRequestFactory, manager: ModelManager) => {
+                let responseItems = List([
+                    new MyModel(1),
+                    new MyModel(12),
+                    new MyModel(123),
+                    new MyModel(1234)
+                ]);
+
+                requestFactory.sent$.subscribe((request: MockRequest) => {
+                    expect(request.method).toBe(RequestMethod.Get);
+                    expect(request.path).toEqual(['test']);
+                    expect(request.query).toEqual({'ids': '1,12,123,1234'});
+
+                    request.respond({
+                        status: 200,
+                        body: itemList(model(MyModel)).encode(responseItems)
+                    })
+
+                });
+
+                manager.loadMany(MyModel, List([1, 12, 123, 1234])).forEach(models => {
+                    expect(models).toEqual(responseItems);
+                })
+            }))
+        );
+
+        it('saving a model with null ID should POST to server', async(inject(
             [RequestFactory, ModelManager],
             (requestFactory: MockRequestFactory, manager: ModelManager) => {
                 requestFactory.sent$.subscribe((request) => {
 
                     expect(request.method).toEqual(RequestMethod.Post);
-                    expect(request.path).toEqual(['test', 'create']);
+                    expect(request.path).toEqual(['test']);
                     expect(request.body).toEqual({kind: 'test::MyModel', id: null});
 
                     let created = new MyModel(42);
@@ -147,11 +183,11 @@ describe('manager.base', () => {
                     expect(saved.id).toEqual(42)
                 });
 
-            })
+            }))
         );
 
 
-        it('saving a model with not-null ID should PUT to server', inject(
+        it('saving a model with not-null ID should PUT to server', async(inject(
             [RequestFactory, ModelManager],
             (requestFactory: MockRequestFactory, manager: ModelManager) => {
                 requestFactory.sent$.subscribe((request) => {
@@ -169,7 +205,30 @@ describe('manager.base', () => {
                 });
 
             })
+        ));
+
+        it('should be possible to resolve the value of a reference property', async(inject(
+            [RequestFactory, ModelManager],
+            (requestFactory: MockRequestFactory, manager: ModelManager) => {
+                let model = new ReferencingModel(12345);
+                expect(model.isResolved('ref')).toBe(false);
+
+                let reference = new MyModel(12345);
+
+                requestFactory.sent$.subscribe((request) => {
+                    expect(request.path).toEqual(['test', '12345']);
+
+                    request.respond({status: 200, body: reference});
+                });
+
+                manager.resolve(model, 'ref').forEach(resolved => {
+                    expect(resolved).toEqual(jasmine.any(ReferencingModel));
+                    expect(resolved.isResolved('ref')).toEqual(true);
+                    expect(resolved.ref).toEqual(reference);
+                });
+            }))
         );
+
     });
 })