+++ CHANGELOG.md
@@ -16,7 +16,7 @@
     Models can now be annotated as `'isAbstract: true'`, to provide better support
     for inheritance between models
 - Can now pass an immutable value to PropertyOptions.default
-- Renamed 'ModelBase.resolveProperty' to 'ModelBase.resolve'
+- Removed 'ManagerBase.resolveProperty' method, moving it to `ModelManager`
 
 
 #### json_codecs
@@ -29,12 +29,12 @@
 #### Manager
 - Renamed ManagerBase to ModelManager, and removed ManagerOptions dependency.
   If methods which extend the manager base are needed, Manager can still be
-  subclassed appropriately.
 - Replaced manager.request module with manager.http module, improving the experience
   for working with managed models.
 - Added ModelManager.save method, which
     - POSTs the model to <managerPath>/create if the id is `null`
     - PUTs the model to <managerPath>/<modelId> if the id is present.
+- Added ModelManager.resolve method, which fetches the value of a given foreign key.
 
 ### Bugfixes
 - Factory created model instances now inherit methods defined on the model